mexing with nosort pixels

+++ _test/test_algorithms/test_cut_data_from_file_job.m
@@ -42,7 +42,8 @@ classdef test_cut_data_from_file_job < TestCase
             assertEqual(dat.s,s);
             assertEqual(dat.e,e);
             assertEqual(pix,pix_ok)
-            assertEqual(unique_runid,1000)
+            % unique_runid usually integer type
+            assertEqual(double(unique_runid),1000)
             npc = accumarray(pix_indx, ones(1,size(pix_indx,1)), [numel(npix),1]);
             assertEqual(npix(:),npc);
         end
@@ -75,7 +76,7 @@ classdef test_cut_data_from_file_job < TestCase
             assertEqual(dat.s,s);
             assertEqual(dat.e,e);
             assertEqual(pix,pix_ok)
-            assertEqual(unique_runid,1000)
+            assertEqual(double(unique_runid),1000)
             npc = accumarray(pix_indx, ones(1,size(pix_indx,1)), [numel(npix),1]);
             assertEqual(npix(:),npc);
         end
@@ -102,7 +103,7 @@ classdef test_cut_data_from_file_job < TestCase
             assertEqual(dat.s,s);
             assertEqual(dat.e,e);
             assertEqual(pix,pix_ok)
-            assertEqual(unique_runid,1000)
+            assertEqual(double(unique_runid),1000)
             npc = accumarray(pix_indx, ones(1,size(pix_indx,1)), [numel(npix),1]);
             assertEqual(npix(:),npc);
         end
